Remove Ethereum document store issuance method
During the last Working Group meeting, the team received feedback that the spec should focus on wrapping for selective redaction, and should not include mention of Ethereum or the document store method, as that would be too technology-specific. This PR has removed these mentions and instead described the document store method as publishing the merkleRoot to a Verifiable Data Registry. 

This PR also includes minor edits to mention that OA has been in production since 2018, and to add TradeTrust as a potential use case.